Fusión de Leaderbaord y FullLeaderboard para evitar código duplicado
