+++ webapp/src/components/Leaderboard.js
@@ -2,10 +2,9 @@ import React, { useEffect, useState, useRef } from "react";
 import {
     Paper, Table, TableBody, TableCell, TableContainer,
     TableHead, TableRow, Typography, CircularProgress,
-    Box, TableSortLabel, Avatar, Button
+    Box, TableSortLabel, Avatar, Button, TablePagination
 } from '@mui/material';
 import Navbar from './Navbar';
-import FullLeaderboard from './FullLeaderboard';
 import { WinRateBar, medalEmojis, medalColors, StickyPlayerHeader, calculatePointsToLevelUp } from './LeaderboardComponents';
 
 const apiEndpoint = process.env.REACT_APP_API_ENDPOINT || 'http://localhost:8000';
@@ -18,6 +17,8 @@ const Leaderboard = () => {
     const [orderBy, setOrderBy] = useState('totalScore');
     const [currentUsername, setCurrentUsername] = useState(null);
     const [showFullLeaderboard, setShowFullLeaderboard] = useState(false);
+    const [page, setPage] = useState(0);
+    const [rowsPerPage, setRowsPerPage] = useState(10);
     const currentUserRef = useRef(null);
     
     // Add a ref to get the Navbar height
@@ -76,31 +77,36 @@ const Leaderboard = () => {
         calculatePointsToLevelUp(sortedPlayers, currentUserIndex, currentUser),
     [sortedPlayers, currentUserIndex, currentUser]);
 
-    // Show full leaderboard component instead if expanded
-    if (showFullLeaderboard) {
-        return (
-            <>
-                <Navbar ref={navbarRef} />
-                {currentUser && (
-                    <StickyPlayerHeader 
-                        player={currentUser} 
-                        rank={currentUserIndex}
-                        pointsToLevelUp={pointsToLevelUp}
-                        navbarHeight={navbarHeight}
-                    />
-                )}
-                <FullLeaderboard 
-                    players={sortedPlayers} 
-                    currentUsername={currentUsername}
-                    order={order}
-                    orderBy={orderBy}
-                    onRequestSort={handleRequestSort}
-                    onCollapseView={() => setShowFullLeaderboard(false)}
-                    navbarHeight={navbarHeight}
-                />
-            </>
-        );
-    }
+    // Calculate which page contains the current user to auto-scroll there in full leaderboard view
+    const userPage = Math.floor(currentUserIndex / rowsPerPage);
+
+    useEffect(() => {
+        // Set the page to show the current user when switching to full leaderboard
+        if (showFullLeaderboard && currentUserIndex >= 0) {
+            setPage(userPage);
+        }
+    }, [currentUserIndex, userPage, showFullLeaderboard]);
+
+    useEffect(() => {
+        // Scroll to current user after rendering
+        if (currentUserRef.current) {
+            currentUserRef.current.scrollIntoView({ behavior: 'smooth', block: 'center' });
+        }
+    }, [page, showFullLeaderboard]);
+
+    const handleChangePage = (event, newPage) => {
+        setPage(newPage);
+    };
+
+    const handleChangeRowsPerPage = (event) => {
+        setRowsPerPage(parseInt(event.target.value, 10));
+        setPage(0);
+    };
+
+    // Toggle between compact and full leaderboard views
+    const toggleLeaderboardView = () => {
+        setShowFullLeaderboard(!showFullLeaderboard);
+    };
 
     if (loading) return (
         <>
@@ -120,6 +126,11 @@ const Leaderboard = () => {
             </Box>
         </>
     );
+
+    // Get the players to display based on the view
+    const displayedPlayers = showFullLeaderboard 
+        ? sortedPlayers.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+        : sortedPlayers.slice(0, 5);
 
     return (
         <>
@@ -142,8 +153,31 @@ const Leaderboard = () => {
                         alignItems: 'center',
                         justifyContent: 'space-between'
                     }}>
-                        <Typography variant="h5" sx={{ fontWeight: 'bold' }}>🏆 Leaderboard</Typography>
-                        <Typography variant="subtitle1">{players.length} Player{players.length !== 1 ? 's' : ''}</Typography>
+                        <Typography variant="h5" sx={{ fontWeight: 'bold' }}>
+                            🏆 {showFullLeaderboard ? 'Full Leaderboard' : 'Leaderboard'}
+                        </Typography>
+                        <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                            <Typography variant="subtitle1" sx={{ mr: 2 }}>
+                                {players.length} Player{players.length !== 1 ? 's' : ''}
+                            </Typography>
+                            {showFullLeaderboard && (
+                                <Button 
+                                    variant="outlined"
+                                    size="small"
+                                    onClick={toggleLeaderboardView}
+                                    sx={{ 
+                                        color: 'white', 
+                                        borderColor: 'white',
+                                        '&:hover': { 
+                                            borderColor: '#e0e0e0',
+                                            bgcolor: 'rgba(255, 255, 255, 0.1)'
+                                        }
+                                    }}
+                                >
+                                    Show Top 5
+                                </Button>
+                            )}
+                        </Box>
                     </Box>
 
                     <TableContainer sx={{ maxHeight: 600 }}>
@@ -168,14 +202,25 @@ const Leaderboard = () => {
                                             </TableSortLabel>
                                         </TableCell>
                                     ))}
+                                    {/* Adding Points To Level Up column only in full view */}
+                                    {showFullLeaderboard && (
+                                        <TableCell align="right">Points To Level Up</TableCell>
+                                    )}
                                 </TableRow>
                             </TableHead>
                             <TableBody>
-                                {/* Top 5 players */}
-                                {sortedPlayers.slice(0, 5).map((player, index) => {
+                                {/* Players list */}
+                                {displayedPlayers.map((player, index) => {
                                     const playerRank = sortedPlayers.findIndex(p => p._id === player._id);
                                     const isCurrentUser = player.username === currentUsername;
                                     const isTop3 = playerRank < 3;
+                                    
+                                    // Calculate points needed for this player to level up
+                                    let pointsNeeded = null;
+                                    if (playerRank > 0) {
+                                        const playerAbove = sortedPlayers[playerRank - 1];
+                                        pointsNeeded = playerAbove.totalScore - player.totalScore + 1;
+                                    }
                                     
                                     let bgColor = 'inherit';
                                     if (isCurrentUser) {
@@ -217,19 +262,34 @@ const Leaderboard = () => {
                                             <TableCell align="right">{player.gamesPlayed}</TableCell>
                                             <TableCell align="right">{player.avgPointsPerGame?.toFixed(2)}</TableCell>
                                             <TableCell align="right"><WinRateBar winRate={player.winRate} /></TableCell>
+                                            {showFullLeaderboard && (
+                                                <TableCell align="right">
+                                                    {playerRank === 0 ? (
+                                                        <Typography sx={{ color: '#66bb6a', fontWeight: 'bold' }}>
+                                                            🏆 Top Player!
+                                                        </Typography>
+                                                    ) : pointsNeeded !== null ? (
+                                                        <Typography>
+                                                            {pointsNeeded}
+                                                        </Typography>
+                                                    ) : '-'}
+                                                </TableCell>
+                                            )}
                                         </TableRow>
                                     );
                                 })}
 
-                                {/* Three dots to expand */}
-                                <TableRow>
-                                    <TableCell colSpan={6} align="center" sx={{ py: 1, cursor: 'pointer' }} onClick={() => setShowFullLeaderboard(true)}>
-                                        <Typography sx={{ color: '#6A5ACD' }}>•••</Typography>
-                                    </TableCell>
-                                </TableRow>
+                                {/* Show expansion dots in compact view */}
+                                {!showFullLeaderboard && (
+                                    <TableRow>
+                                        <TableCell colSpan={6} align="center" sx={{ py: 1, cursor: 'pointer' }} onClick={toggleLeaderboardView}>
+                                            <Typography sx={{ color: '#6A5ACD' }}>•••</Typography>
+                                        </TableCell>
+                                    </TableRow>
+                                )}
 
-                                {/* User context if not in top 5 */}
-                                {!isCurrentUserTop5 && currentUser && (
+                                {/* User context if not in top 5 (only in compact view) */}
+                                {!showFullLeaderboard && !isCurrentUserTop5 && currentUser && (
                                     <>
                                         {/* Show player above if exists */}
                                         {currentUserIndex > 0 && (
@@ -321,21 +381,36 @@ const Leaderboard = () => {
                             </TableBody>
                         </Table>
                     </TableContainer>
+
+                    {/* Pagination only for full leaderboard view */}
+                    {showFullLeaderboard && (
+                        <TablePagination
+                            rowsPerPageOptions={[10, 25, 50]}
+                            component="div"
+                            count={sortedPlayers.length}
+                            rowsPerPage={rowsPerPage}
+                            page={page}
+                            onPageChange={handleChangePage}
+                            onRowsPerPageChange={handleChangeRowsPerPage}
+                        />
+                    )}
                 </Paper>
                 
-                {/* Button to show full leaderboard */}
-                <Box sx={{ display: 'flex', justifyContent: 'center', mt: 2 }}>
-                    <Button 
-                        variant="contained" 
-                        onClick={() => setShowFullLeaderboard(true)}
-                        sx={{ 
-                            bgcolor: '#3949AB',
-                            '&:hover': { bgcolor: '#303F9F' }
-                        }}
-                    >
-                        Show Full Leaderboard
-                    </Button>
-                </Box>
+                {/* Button to toggle leaderboard view */}
+                {!showFullLeaderboard && (
+                    <Box sx={{ display: 'flex', justifyContent: 'center', mt: 2 }}>
+                        <Button 
+                            variant="contained" 
+                            onClick={toggleLeaderboardView}
+                            sx={{ 
+                                bgcolor: '#3949AB',
+                                '&:hover': { bgcolor: '#303F9F' }
+                            }}
+                        >
+                            Show Full Leaderboard
+                        </Button>
+                    </Box>
+                )}
             </Box>
         </>
     );
